refactor: organize resource page layout and separate course logic
