+++ frontend/src/components/CertificateSection.jsx
@@ -3,35 +3,7 @@ import { useState } from "react";
 import { CategoryFilter } from "@/components/filter";
 import GradientHeading from "@/components/Text/GradientHeading";
 import { Loader2 } from "lucide-react";
-
-const certTags = ["All", "Beginner", "Intermediate", "Professional", "Free", "React"];
-const certificates = [
-  {
-    title: "Google UX Design Certificate",
-    platform: "Coursera",
-    tags: ["Beginner", "Free", "Design", "Certificate", "Official"],
-    description: "Learn the foundations of UX/UI design and research.",
-    url: "https://www.coursera.org/professional-certificates/google-ux-design",
-    icon: "https://www.gstatic.com/images/icons/material/system/1x/design_services_black_24dp.png"
-  },
-  {
-    title: "Meta Front-End Developer",
-    platform: "Coursera",
-    tags: ["Intermediate", "React", "Frontend", "Certificate", "Official"],
-    description: "Build front-end apps with HTML, CSS, JS, and React.",
-    url: "https://www.coursera.org/professional-certificates/meta-front-end-developer",
-    icon: "https://upload.wikimedia.org/wikipedia/commons/0/05/Facebook_Logo_%282019%29.png"
-  },
-  {
-    title: "IBM Full Stack Developer",
-    platform: "Coursera",
-    tags: ["Intermediate", "Fullstack", "Paid", "Certificate"],
-    description: "Master web dev, React, Node.js, and cloud-native tools.",
-    url: "https://www.coursera.org/professional-certificates/ibm-full-stack-cloud-developer",
-    icon: "https://upload.wikimedia.org/wikipedia/commons/5/51/IBM_logo.svg"
-  }
-];
-
+import { certTags, certificates } from "../data/certificates";
 
 export function CertificateSection() {
   const [selectedTag, setSelectedTag] = useState("All");
