+++ frontend/src/data/docs.js
@@ -1,0 +1,53 @@
+export const docCategories = [
+  "All",
+  "Docs",
+  "Tools",
+  "React",
+  "CSS",
+  "Design",
+  "Official",
+];
+export const docs = [
+  {
+    title: "React",
+    description: "Build UI with components using the React library.",
+    url: "https://react.dev/",
+    icon: "https://raw.githubusercontent.com/devicons/devicon/master/icons/react/react-original.svg",
+    tags: ["React", "Frontend", "Docs", "Official"],
+  },
+  {
+    title: "TailwindCSS",
+    description: "Utility-first CSS framework for rapid UI building.",
+    url: "https://tailwindcss.com/docs",
+    icon: "https://raw.githubusercontent.com/devicons/devicon/master/icons/tailwindcss/tailwindcss-plain.svg",
+    tags: ["CSS", "Frontend", "Docs", "Official"],
+  },
+  {
+    title: "TypeScript",
+    description: "JavaScript with types. Strong tooling and safety.",
+    url: "https://www.typescriptlang.org/docs/",
+    icon: "https://raw.githubusercontent.com/devicons/devicon/master/icons/typescript/typescript-original.svg",
+    tags: ["Frontend", "Backend", "Docs", "Official"],
+  },
+  {
+    title: "Shadcn UI",
+    description: "Accessible, composable components built with Tailwind.",
+    url: "https://ui.shadcn.dev/",
+    icon: "https://ui.shadcn.dev/favicon.ico",
+    tags: ["UI", "Design", "Tool", "Popular"],
+  },
+  {
+    title: "GitHub Docs",
+    description: "Learn Git, version control, and GitHub best practices.",
+    url: "https://docs.github.com/",
+    icon: "https://github.githubassets.com/favicons/favicon.svg",
+    tags: ["Tool", "Docs", "Official"],
+  },
+  {
+    title: "Next.js",
+    description: "React framework for building fullstack web apps.",
+    url: "https://nextjs.org/docs",
+    icon: "https://nextjs.org/static/favicon/favicon.ico",
+    tags: ["React", "Fullstack", "Docs", "Popular"],
+  },
+];
